+++ Drakhtar/Controllers/PlayerController.h
@@ -11,7 +11,7 @@ class PlayerController final : public UnitsController {
   Button* skipTurnButton_ = nullptr;
   bool tutorialDone_ = false;
   std::vector<Vector2D<double>> pathToRoute(
-      const std::vector<Vector2D<byte>>& path) const;
+      const std::vector<Vector2D<uint16_t>>& path) const;
 
  public:
   PlayerController(Board* board, GameScene* scene, Team* team,
