+++ Drakhtar/GameObjects/Board.cpp
@@ -6,7 +6,7 @@
 #include "Utils/Constants.h"
 #include "third_party/AStar.h"
 
-Board::Board(Scene* scene, const byte rows, const byte columns,
+Board::Board(Scene* scene, const uint16_t rows, const uint16_t columns,
              const double cellSize)
     : GameObject(scene, nullptr, Vector2D<int>(0, 0), Vector2D<int>(0, 0)),
       rows_(rows),
@@ -19,14 +19,14 @@ Board::Board(Scene* scene, const byte rows, const byte columns,
                 static_cast<int>((WIN_HEIGHT - size_.getY()) / 2));
 
   // Fills the board with empty boxes
-  for (byte x = 0; x < columns_; x++) {
-    for (byte y = 0; y < rows_; y++) {
+  for (uint16_t x = 0; x < columns_; x++) {
+    for (uint16_t y = 0; y < rows_; y++) {
       auto pos =
           Vector2D<int>(static_cast<int>(position_.getX() + x * cellSize_),
                         static_cast<int>(position_.getY() + y * cellSize_));
       const auto size = Vector2D<int>(static_cast<int>(cellSize_),
                                       static_cast<int>(cellSize_));
-      const auto box = new Box(scene, pos, size, Vector2D<byte>(x, y));
+      const auto box = new Box(scene, pos, size, Vector2D<uint16_t>(x, y));
       addChild(box);
     }
   }
@@ -38,6 +38,6 @@ SDL_Rect Board::getRect() const {
   return {position_.getX(), position_.getY(), size_.getX(), size_.getY()};
 }
 
-Box* Board::getBoxAt(const byte x, const byte y) const {
+Box* Board::getBoxAt(const uint16_t x, const uint16_t y) const {
   return reinterpret_cast<Box*>(children_[x * rows_ + y]);
 }
